fixup! fix(ivy): ngcc - prefer JavaScript source files when resolving module imports
